# BUG - 0 quantity deletes product from cart on product page and browse page. Should only happen on update or single add to cart button child list
git-svn-id: https://dev.virtuemart.net/svn/virtuemart@1716 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ branches/virtuemart-1.1/virtuemart/classes/ps_cart.php
@@ -342,7 +342,7 @@ class vm_ps_cart {
 		}
 		$deleted_prod = 0;
 		$updated_prod = 0;
-		if ($quantity == 0) {
+		if ($quantity == 0 && strtolower($func) == "cartupdate") {
 			$deleted_prod = $this->delete($d);
 		}
 		else {
